requirements.txt
datasets
spacy
wordfreq
lemminflect
pyyaml